refactor/Service: focus on being easier to do unit test

+++ src/main/java/com/example/demo/service/LaboratorioService.java
@@ -2,8 +2,9 @@ package com.example.demo.service;
 
 import com.example.demo.entity.Laboratorio;
 import com.example.demo.repository.LaboratorioRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,8 +12,11 @@ import java.util.Optional;
 @Service
 public class LaboratorioService {
 
-    @Autowired
-    private LaboratorioRepository laboratorioRepository;
+    private final LaboratorioRepository laboratorioRepository;
+
+    public LaboratorioService(LaboratorioRepository laboratorioRepository) {
+        this.laboratorioRepository = laboratorioRepository;
+    }
 
     public List<Laboratorio> getAllLaboratorios() {
         return laboratorioRepository.findAll();
@@ -27,16 +31,18 @@ public class LaboratorioService {
     }
 
     public Laboratorio updateLaboratorio(Long id, Laboratorio laboratorioDetails) {
-        Optional<Laboratorio> optionalLaboratorio = laboratorioRepository.findById(id);
-        if (optionalLaboratorio.isPresent()) {
-            Laboratorio laboratorio = optionalLaboratorio.get();
-            laboratorio.setNome(laboratorioDetails.getNome());
-            return laboratorioRepository.save(laboratorio);
-        }
-        return null;
+        return laboratorioRepository.findById(id)
+                .map(laboratorio -> {
+                    laboratorio.setNome(laboratorioDetails.getNome());
+                    return laboratorioRepository.save(laboratorio);
+                })
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Laboratório não encontrado"));
     }
 
     public void deleteLaboratorio(Long id) {
+        if (!laboratorioRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Laboratório não encontrado");
+        }
         laboratorioRepository.deleteById(id);
     }
 }
